pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tsd"
version = "0.1.0"
description = "Text spotting dataset tools"
authors = [
    { name = "David Jones", email = "d.t.jones@outlook.com" },
]
maintainers = [
    { name = "David Jones", email = "d.t.jones@outlook.com" },
]
license = { file="LICENSE" }
readme = "README.md"
keywords = ["text", "spotting", "recognition", "detection"]
dependencies = [
    "numpy>1.26.0",
    "Pillow>10.0.0",
    "shapely>2.0.0"
]

[project.urls]
homepage = "https://github.com/davidtjones/tsd"
repository = "https://github.com/davidtjones/tsd"